app/backend/ecg/service.py
# backend/ecg/service.py
from typing import List, Dict
import numpy as np
from fastapi import UploadFile
from . import ecg_classification_api  # Import the existing API
import io
from PIL import Image

# Assuming ecg_classification_api has a classify_ecg function:
# from .ecg_classification_api import classify_ecg
# If it's a class, you might instantiate it here:
# ecg_classifier = ecg_classification_api.ECGClassifier()

async def classify_ecg_signal(ecg_signal: List[float]) -> Dict:
    """Classify a list of ECG data points using the existing API."""
    # Convert list to numpy array if needed by your existing API
    ecg_signal_np = np.array(ecg_signal)

    # Assuming classify_ecg in ecg_classification_api takes a numpy array
    classification_result = ecg_classification_api.classify_ecg(ecg_signal_np) 

    return classification_result

async def classify_ecg_image(image_file: UploadFile) -> Dict:
    """Classify an ECG image using the existing API."""
    try:
        contents = await image_file.read()
        image = Image.open(io.BytesIO(contents))
        classification_result = ecg_classification_api.classify_ecg_from_image(image)
        return classification_result
    except Exception as e:
        return {"error": f"Error processing image: {e}"}


# Add other ECG-related operations as needed, e.g.,
# - Preprocessing
# - Feature extraction
# - Model loading/management
# - Data storage/retrieval (if not handled by a separate DB layer)

# Example if you need to load a model in this service:
# def load_ecg_model():
#     #  Implementation here.  Make sure path is correct relative to this file.
#     pass