app/backend/requirements.txt
fastapi
uvicorn[standard]
motor
PyJWT
passlib[bcrypt]
python-multipart
python-jose[cryptography]
bcrypt
pydantic>=2.0.0
pydantic-settings
werkzeug
numpy
opencv-python
gunicorn
Pillow
pydicom
pylibjpeg
pylibjpeg-libjpeg
python-gdcm
pydicom
pylibjpeg
PyWavelets
torch
torchvision
timm